Refactor connectors english-campaign-rb - use constructor injection instead of field injection
